pyproject.toml
[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

[project]
name = "notion-caldav-sync"
version = "0.1.0"
description = "Notion → iCloud Calendar sync worker"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
dependencies = [
  "caldav>=1.3.7",
  "icalendar>=6.1.0",
  "python-dateutil>=2.9.0",
  "httpx>=0.27.0",
  "requests>=2.32.0",
]

[dependency-groups]
dev = [
  "python-dotenv>=1.0",
  "typer>=0.12",
  "pytest>=8.3",
  "pytest-asyncio>=0.23",
  "ruff>=0.5",
  "workers-py>=1.7",
]

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[tool.hatch.build.targets.sdist]
include = ["src/app", "README.md", "pyproject.toml", "LICENSE"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
markers = [
  "integration: marks tests that talk to Notion/iCloud APIs",
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
